Split testqgsprocessingalgs.cpp into two parts to avoid timeouts on CI
